+++ repair_binance_data.py
@@ -1,1 +1,2 @@
 # todo: change settings to default, delete all data, then run monthly and run daily
+# todo: set "daily_hist_complete = 0 AND "  "monthly_hist_complete = 0 AND "
